Moved main.c into source folder
