+++ src/Makefile
@@ -4,7 +4,7 @@ all: main
 
 main: main.c
 	@echo "Compliling HTTP server."
-	${CC} webpage.c Server.c main.c -l sqlite3 -o main
+	${CC} server/webpage.c server/Server.c main.c -l sqlite3 -o main
 
 clean:
 	@echo "Cleaning up..."
