+++ src/server/index.html
@@ -52,9 +52,9 @@
                 // create a data set
                 var data = anychart.data.set([
 			//INSERT//
-["2022-04-26","1","2","#ab0000", "#009933"],
-["2022-04-27","0","3","#ab0000", "#009933"],
-["2022-04-28","3","0","#ab0000", "#009933"],
+			["2022-04-26","1","2","#ab0000", "#009933"],
+			["2022-04-27","0","3","#ab0000", "#009933"],
+			["2022-04-28","3","0","#ab0000", "#009933"],
                 ]);
 
                 // map the data
