+++ src/server/webpage.c
@@ -4,18 +4,16 @@ static int callback(void *data, int argc, char **argv, char **azColName)
 {
 	char connection_data[1024] = {'\0'};
 
-	strcat(connection_data, "[");
+	strcat(connection_data, "\t\t\t[");
 
 	for(int i = 0; i< argc; i++){
-		//printf("%s = %s\n", azColName[i], argv[i] ? argv[i] : "NULL");
 		char tmp[1024] = {'\0'};
 		sprintf(tmp, "\"%s\",", argv[i] ? argv[i] : "NULL");
 		strcat(connection_data, tmp);
 	}
    
 	strcat(connection_data, "\"#ab0000\", \"#009933\"],\n");
-	//printf("\"#ab0000\", \"#009933\"],\n");
-	printf("%s\n", connection_data);
+	printf("%s", connection_data);
 
 	FILE* f = fopen("connection.dat", "a");
 
@@ -30,7 +28,7 @@ static int callback(void *data, int argc, char **argv, char **azColName)
 	return 0;
 }
 
-int update_webpage(const char* file_path) 
+int update_webpage(const char* file_source,const char* file_destination) 
 {
 	sqlite3 *db;
 	char *zErrMsg = 0;
@@ -39,7 +37,7 @@ int update_webpage(const char* file_path)
 	const char* data = "Callback function called";
 
 	/* Open database */
-	rc = sqlite3_open("../database/connection.db", &db);
+	rc = sqlite3_open("database/connection.db", &db);
    
 	if( rc ) 
 	{
@@ -58,13 +56,13 @@ int update_webpage(const char* file_path)
     	FILE *html_data;
 	
 	// Open file with HTML data
-    	html_data = fopen("template.html", "r");
+    	html_data = fopen(file_source, "r");
     	while (fgets(line, sizeof(line), html_data))
     	{
         	strcat(response_data, line);
 
 		char* check;
-		check = strstr(line, "//INSERT//");
+		check = strstr(line, "Connection Data:");
 		if(check) 
 		{
 			printf("%s\n", line);
@@ -107,7 +105,7 @@ int update_webpage(const char* file_path)
         	strcat(response_data, line);
     	}
 
-	FILE* index = fopen("index.html", "w");
+	FILE* index = fopen(file_destination, "w");
 
 	if(index == NULL) {
 		printf("Error: Could not create file.");
