+++ src/server/webpage.h
@@ -9,6 +9,6 @@
 // Retrieves the host up to date HTML page from the file path
 void retrieve_webpage(const char* file_path, char* buffer, int buffer_size);
 
-int update_webpage(const char* file_path);
+int update_webpage(const char* file_source, const char* file_destination);
 
 #endif
